refactor: extract SocialLoginService to interface
