+++ src/main/java/dev/smjeon/commerce/oauth/application/KakaoLoginService.java
@@ -1,0 +1,95 @@
+package dev.smjeon.commerce.oauth.application;
+
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import dev.smjeon.commerce.oauth.SocialUserInfo;
+import dev.smjeon.commerce.oauth.domain.SocialLoginUser;
+import dev.smjeon.commerce.oauth.domain.SocialUserRepository;
+import dev.smjeon.commerce.oauth.kakao.KakaoConfig;
+import dev.smjeon.commerce.oauth.kakao.dto.KakaoToken;
+import dev.smjeon.commerce.user.domain.UserRole;
+import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
+import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.BodyInserters;
+import org.springframework.web.reactive.function.client.WebClient;
+
+@Service
+public class KakaoLoginService implements SocialLoginService {
+    private KakaoConfig kakaoConfig;
+    private SocialUserRepository socialUserRepository;
+    private ModelMapper modelMapper;
+
+    public KakaoLoginService(KakaoConfig kakaoConfig, SocialUserRepository socialUserRepository, ModelMapper modelMapper) {
+        this.kakaoConfig = kakaoConfig;
+        this.socialUserRepository = socialUserRepository;
+        this.modelMapper = modelMapper;
+    }
+
+    @Override
+    public String getRedirectUrl() {
+        StringBuilder redirectUrl = new StringBuilder();
+        redirectUrl.append(kakaoConfig.getAuthorizationUrl());
+        redirectUrl.append("?client_id=").append(kakaoConfig.getClientId());
+        redirectUrl.append("&redirect_uri=").append(kakaoConfig.getRedirectUrl());
+        redirectUrl.append("&response_type=code");
+
+        return redirectUrl.toString();
+    }
+
+    @Override
+    public String getAccessToken(String code) {
+        WebClient webClient = WebClient.builder()
+                .baseUrl(kakaoConfig.getAccessTokenUrl())
+                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                .defaultHeader(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+                .build();
+
+        String body = webClient.post()
+                .body(BodyInserters.fromFormData("grant_type", "authorization_code")
+                        .with("client_id", kakaoConfig.getClientId())
+                        .with("redirect_uri", kakaoConfig.getRedirectUrl())
+                        .with("client_secret", kakaoConfig.getClientSecret())
+                        .with("code", code))
+                .retrieve()
+                .bodyToMono(String.class)
+                .block();
+
+        KakaoToken token = new Gson().fromJson(body, KakaoToken.class);
+
+        return token.getAccessToken();
+    }
+
+    @Override
+    public SocialUserInfo getUserInfo(String accessToken) {
+        WebClient webClient = WebClient.builder()
+                .baseUrl(kakaoConfig.getUserInfoUrl())
+                .defaultHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+                .defaultHeader(HttpHeaders.ACCEPT, MediaType.APPLICATION_JSON_VALUE)
+                .defaultHeader(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
+                .build();
+
+        String body = webClient.post()
+                .retrieve()
+                .bodyToMono(String.class)
+                .block();
+
+        String kakaoId = new JsonParser().parse(body).getAsJsonObject().get("id").getAsString();
+        JsonObject properties = new JsonParser().parse(body).getAsJsonObject().get("properties").getAsJsonObject();
+        String nickName = properties.getAsJsonObject().get("nickname").getAsString();
+        JsonObject kakaoAccount = new JsonParser().parse(body).getAsJsonObject().get("kakao_account").getAsJsonObject();
+        String email = kakaoAccount.getAsJsonObject().get("email").getAsString();
+
+        SocialUserInfo userInfo = new SocialUserInfo(kakaoId, nickName, email);
+        return userInfo;
+    }
+
+    public SocialLoginUser save(SocialUserInfo userInfo) {
+        SocialLoginUser user = modelMapper.map(userInfo, SocialLoginUser.class);
+        user.updateUserRole(UserRole.BUYER);
+
+        return socialUserRepository.save(user);
+    }
+}
